Refactor duplicate match job
This refactors the job which searches for a duplicate match to simplify
the code and make it easier to read.
